Manager test completed

+++ src/main/java/staff/Employee.java
@@ -6,7 +6,7 @@ public abstract class Employee {
     private String NInumber;
     private double salary;
 
-    public Employee(String name, String NInumber, int salary){
+    public Employee(String name, String NInumber, double salary){
         this.name = name;
         this.NInumber = NInumber;
         this.salary = salary;
