+++ src/main/java/staff/management/Manager.java
@@ -6,7 +6,7 @@ public class Manager extends Employee {
     
     private String deptName;
 
-    public Manager(String name, String NInumber, int salary, String deptName) {
+    public Manager(String name, String NInumber, double salary, String deptName) {
         super(name, NInumber, salary);
         this.deptName = deptName;
     }
